Clear every clippy warning and require that they stay cleared
Ninety-six were mechanical and clippy fixed them. The rest were read one at
a time, because a lint that fires on deliberate code is a question, not an
instruction.

Most named something worth changing. A `drop` on a value implementing no
`Drop` was ending a borrow while reading as a release. Nine return types
wanted names and now have them. Three constant assertions moved to compile
time. An `Arc` around a type that is neither `Send` nor `Sync` became an
`Rc`, which the compiler confirmed by accepting it. A cursor loop cloned
the vector it walked because the body mutated it; it now takes the stale
set first, which is both cheaper and says why. A `Result<_, ()>` names its
one failure. Two large enum variants are boxed so their neighbours in the
same queue are not sized by them.

Two thresholds are recorded in clippy.toml with reasons rather than
contorting the code to fit defaults tuned for other kinds of program. Draw
calls and transport constructors take genuinely independent parameters, and
grouping them would invent structs whose only purpose is lowering a count.
Several senders return the value they could not send so a caller can
recover it, which is the design and not an oversight; boxing those would
allocate on the submit path and hide that the error is the value. The two
largest of those say so at the site instead, since one general threshold
should not swallow a 520-byte case.

clippy resolves clippy.toml from the crate being linted rather than walking
up to the workspace root, so a plain workspace run silently used the
defaults. The check now points it at the root, which is what makes the
recorded thresholds apply at all.

The style guide now requires the gate to end with no warnings, and says
which of the three resolutions to reach for. A hundred standing warnings is
not a hundred small debts; it is a filter that hides the next real one.

+++ crates/sophia-backend-live/src/scanout/rendered_scanout/exporter/worker/damage.rs
@@ -158,8 +158,8 @@ fn slot_repaint_table(
                         .map(|rect| sophia_renderer_live::NativeCompositionDamageRect {
                             x: rect.x,
                             y: rect.y,
-                            width: i32::from(rect.width),
-                            height: i32::from(rect.height),
+                            width: rect.width,
+                            height: rect.height,
                         })
                         .collect(),
                 ));
